+++ docs/src/content/utilities/_meta.js
@@ -1,4 +1,6 @@
 export default {
+  bpjs: 'BPJS',
+  mock: 'Mock',
   nlp: 'NLP',
   pdp: 'PDP (Privacy)',
   currency: 'Currency',